[Feat] add, remove, accept, friends backend done

+++ app/back-end/dashboards/views.py
@@ -77,10 +77,11 @@ class RemoveFriendshipView(APIView):
             user_remove = User.objects.get(username=username)
         except User.DoesNotExist:
             return Response({'error': 'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
-
         try:
             friendship = Friendship.objects.get(Q(user_from=user_from, user_to=user_remove)|
                                                 Q(user_from=user_remove, user_to=user_from))
+            if friendship.u_one_is_blocked_u_two == True or friendship.u_two_is_blocked_u_one == True:
+                return Response({'error': 'Friendship blocked'}, status=status.HTTP_400_BAD_REQUEST)
             friendship.delete()
             return Response({'success': 'Friendship removed'}, status=status.HTTP_200_OK)
         except Friendship.DoesNotExist:
@@ -93,38 +94,44 @@ class AcceptFriendshipView(APIView):
     def post(self, request):
         user_from = request.user
         username = request.data.get('username')
-
         try:
-            user_remove = User.objects.get(username=username)
+            user_accept = User.objects.get(username=username)
         except User.DoesNotExist:
             return Response({'error': 'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
-
         try:
-            friendship = Friendship.objects.get(Q(user_from=user_from, user_to=user_remove)|
-                                                Q(user_from=user_remove, user_to=user_from))
+            friendship = Friendship.objects.get(Q(user_from=user_from, user_to=user_accept)|
+                                                Q(user_from=user_accept, user_to=user_from))
+            if friendship.is_accepted == True:
+                return Response({'error': 'Friendship already accepted'}, status=status.HTTP_400_BAD_REQUEST)
+            elif friendship.u_one_is_blocked_u_two == True or friendship.u_two_is_blocked_u_one == True:
+                return Response({'error': 'Friendship blocked'}, status=status.HTTP_400_BAD_REQUEST)
             friendship.is_accepted = True;
             friendship.save()
             return Response({'success': 'Friendship accepted'}, status=status.HTTP_200_OK)
         except Friendship.DoesNotExist:
-            return Response({'error': 'Friendship does not accepted'}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'error': 'Friendship does not exist'}, status=status.HTTP_400_BAD_REQUEST)
 
-# class AddFriendshipView(APIView):
-#     authentication_classes = [JWTAuthentication]
-#     permission_classes = [IsAuthenticated]
+class AddFriendshipView(APIView):
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsAuthenticated]
 
-#     def post(self, request):
-#         user_from = request.user
-#         username = request.data.get('username')
+    def post(self, request):
+        user_from = request.user
+        username = request.data.get('username')
+        try:
+            user_add = User.objects.get(username=username)
+        except User.DoesNotExist:
+            return Response({'error': 'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+        if user_from.id == user_add.id:
+            return Response({'error': 'wach nta wa7id t ajoute rassk?'}, status=status.HTTP_400_BAD_REQUEST)
+        elif Friendship.objects.filter(Q(user_from=user_from, user_to=user_add) |
+                                     Q(user_from=user_add, user_to=user_from)).exists():
+            return Response({'error': 'Friendship alrady exist'}, status=status.HTTP_400_BAD_REQUEST)
+        try:
 
-#         try:
-#             user_remove = User.objects.get(username=username)
-#         except User.DoesNotExist:
-#             return Response({'error': 'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
-
-#         try:
-#             friendship = Friendship.objects.get(Q(user_from=user_from, user_to=user_remove)|
-#                                                 Q(user_from=user_remove, user_to=user_from))
-#             friendship.delete()
-#             return Response({'success': 'Friendship removed'}, status=status.HTTP_200_OK)
-#         except Friendship.DoesNotExist:
-#             return Response({'error': 'Friendship does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+            friendship = Friendship.objects.create(user_from=user_from, user_to=user_add, 
+                                                   is_accepted = True)
+            friendship.save()
+            return Response({'success': 'Friendship Added'}, status=status.HTTP_200_OK)
+        except Friendship.DoesNotExist:
+            return Response({'error': 'Friendship does not exist'}, status=status.HTTP_400_BAD_REQUEST)
